- Consolidate quoting logic into formatting - Allow 'cl info -f args' to include all request_* fields
